acrn-uefi-sign: add class to sign acrn.efi
uefi-sign from meta-intel assumes an image recipe, but in acrn-hypervisor we
deploy in a non-standard way. Refactor the code to work for this use-case.

Use DISTRO_FEATURES "secureboot" to conditionally enable, since we need this
in recipe space (globally) not just image space (IMAGE_FEATURES)

Reuse the following variables from uefi-sign:
* SECURE_BOOT_SIGNING_CERT_HASH
* SECURE_BOOT_SIGNING_KEY_HASH

Introduce our own variables for the following to not collide with uefi-sign:
* ACRN_SIGNING_BINARIES
* ACRN_SIGNING_DIR
* ACRN_SIGN_BEFORE
* ACRN_SIGN_AFTER

Add weak (??=) sensible defaults.

Signed-off-by: Tim Orling <timothy.t.orling@linux.intel.com>