distro: enable DISTRO_EXTRA_FEATURES
With the addition of 'secureboot' support, we need to be
able to conditionally set DISTRO_FEATURES for SOS and UOS.

Introduce DISTRO_EXTRA_FEATURES_SOS to add features only
to sos and DISTRO_EXTRA_FEATURES_UOS to add features only
to uos.

In the common acrn-demo.inc file, these both resolve to
the more general DISTRO_EXTRA_FEATURES, so that the common
file can remain universal.

Signed-off-by: Tim Orling <timothy.t.orling@linux.intel.com>