acrn-common.inc: fixes and debug for nuc7i7DNB sdc
Add /dev/sdb3 as a rootfs option, as this is what a USB flash drive
will resolve to when there is an internal /dev/sda

Change second UART from ttyS4 to ttyS1. This might be wrong
if the context is VM_SOS_COM2 for vUART for the guest.

Change logging to verbose to help diagnose Secure Boot hang

Signed-off-by: Tim Orling <timothy.t.orling@linux.intel.com>